// src/models/dtos/book.dto.ts
export class BookDto {
    bookId: string;
    bookName: string;
    version: string;
    pageCount: string;
    s3PdfURL: string
    teacherId: string;
    teacherName: string;
    subjectId: string;
    subjectName: string;
    classId: string;
    className: string;
    instituteId: string;
    instituteName: string;
    stateId: string;
    stateName: string;
    createDate: string;
}